Always link to bundled hiredis
Ruby's mkmf cannot be configured to use a local, static library instead
of a global dynamic one. It tried to link against a version that was
already installed in /usr/lib or similar, which broke when hiredis was
upgraded to 0.10 (a few function symbols changed). This change makes
sure the extension is always linked against the bundled, static version.