Fix memory leak for MRI >= 1.9.2
The rb_fdset_t API was introduced in 1.9.2 as an alternative to using a
POSIX fd_set (see select(2)). It was created to support systems that
allow passing file descriptors higher than FD_SETSIZE to select(2) (also
see [1]).

To do this, the file descriptor bitmap is dynamically allocated and
resized as they are set. This means that unlike the POSIX fd_set, it
also needs to be freed when done. I missed this in f85e8c65.

[1]: https://github.com/ruby/ruby/blob/ruby_1_9_2/thread.c#L2228-L2253