Don't attempt to build the C extension on unsupported platforms
hiredis doesn't support Windows at the moment. It is possible to get it
to compile in some cygwin environments, but that may not be the
environment the user is on. This attempts to detect if hiredis would
build successfully, and skips building the native extension on those
platforms. This means that hiredis will fall back to the slower
implementation. This is fine for developers getting started, and if
speed is a requirement, the user can modify their build environment to
build the C extension.

While this case likely means Windows, the implementation doesn't care,
and skips all environments without `socket.h`