Add support for Ruby 3.0 Fiber Scheduler
This implements support for the Ruby 3.0+ Fiber Scheduler interface.

If a Fiber Scheduler (such as the one defined by the Async library)
is present, then we invoke its hooks whenever we're about to perform
an IO wait operation. This makes it possible for the scheduler to
pass control to another Fiber/task while waiting for the socket to
become readable/writable.