Migrate CI to GitHub action (#75)
* CI: Migrate CI to GitHub Actions

The building on travis-ci.org is ceased on June 15th, 2021. So we need to
migrate CI to other services.
There are some CI services, but I recommend GitHub Actions. Because it doesn't
require another account and supports multiples OSs.

* Fix an error message when using Redis with Docker

* Bump `rake`

Without this, exist-status will be 0 even if the tests failed.

* Skip tests that using `#flush`

`#flush` doesn't work on hi-redis 1.0.
I think this is related to https://github.com/redis/hiredis/issues/975.